- Frontend improvements (Eugene)
git-svn-id: svn://svn.zabbix.com/trunk@2590 97f52cf1-0a1b-0410-bd0e-c28be96e8082

+++ frontends/php/index.php
@@ -79,7 +79,7 @@
 	}
 
 
-	if(isset($reconnect))
+	if(isset($reconnect) && isset($sessionid))
 	{
 		$sql="delete from sessions where sessionid='$sessionid'";
 		DBexecute($sql);
